Moved questionnaire views to a seperate file, changed teamDetail objects to Users
